BUG: Avoid copy to folder itself.

+++ ea_dataset_import.m
@@ -158,12 +158,12 @@ else
             end
 
             % copy DICOM data to sourcedata:
-            %without DICOM, error when subject only has DICOMS inside
-            if ~exist(fullfile(dest_dir_subj, 'sourcedata', subjID,'DICOM'), 'dir')
+            % without DICOM, error when subject only has DICOMS inside
+            if ~strcmp(source_dir{subj_idx}, fullfile(dest_dir_subj, 'sourcedata', subjID)) ...
+                    && ~exist(fullfile(dest_dir_subj, 'sourcedata', subjID, 'DICOM'), 'dir')
                 copyfile(fullfile(source_dir{subj_idx}, 'DICOM'), fullfile(dest_dir_subj, 'sourcedata', subjID));
             end
             subj_ids(subj_idx).name =char(subjID);
-
         end
 
         % directory definitions for conversion
